requirements.txt
pandas==2.2.3
plotly==6.0.1
python_dateutil==2.9.0.post0
streamlit==1.44.0
